Add Waveform map sketch
I wondered what a waveform would look like if you colored each sample
based on it's amplitude. The result was interesting and helped me learn
about Float32Arrays and drag and drop in the browser.

Maybe the waveform itself shouldn't be represented but instead maybe
the frequencies in it as they correspond to a molecular spectrometer?