Create linear gradient between two colors for triangles' fill
Playing with the idea of interpolating between two colors in js.

Broke down the color into it's RGB components, took the difference
between each component of the two color, multiplied it by the value
between 0 to 1 and added it to the component of the first color.
Finally covert back to hex.

One important detail was the flooring of the components decimal number
before converting to hex. This prevented it from changing the columns
of the other columns. IE, red's decimal place couldn't change the value
of green or blue.