feat: Use quadtree to get neighbor nodes to push against
Tried multiple quadtree implementations and the fastest and most
straight forward to implement was the d3-quadtree.

They recommend in the README.md to destroy and recreate the quadtree
every frame if you are changing multiple points. Seems counter intuitive
but maybe it works? Also by destroying, they mean clearing the quadtree
but retaining the object. This is a potential upgrade.